refactor(originations): encapsulate explorer link and origination check logic
