+++ src/hooks/useExplorerOperationLink.js
@@ -1,0 +1,17 @@
+import { useMemo } from 'react';
+
+const explorerNetworks = {
+  florence: 'florencenet',
+};
+
+const useExplorerOperationLink = (transactionHash) => {
+  const explorerOperationLink = useMemo(() => {
+    return `https://${
+      explorerNetworks[process.env.REACT_APP_TEZOS_NETWORK]
+    }.tzkt.io/${transactionHash}`;
+  }, [transactionHash]);
+
+  return { explorerOperationLink };
+};
+
+export default useExplorerOperationLink;
